refactor: bonusPts 관련
